backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
python-dotenv==1.0.0
google-genai>=1.0.0
replicate==0.23.1
pillow==10.2.0
aiohttp>=3.10.0
httpx==0.26.0
